+++ lib/ui/academics/AcademicsHomePanel.dart
@@ -26,6 +26,7 @@ import 'package:illinois/service/Guide.dart';
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/academics/AcademicsAppointmentsContentWidget.dart';
 import 'package:illinois/ui/academics/AcademicsEventsContentWidget.dart';
+import 'package:illinois/ui/academics/EssentialSkillsCoachDashboardPanel.dart';
 import 'package:illinois/ui/academics/MedicineCoursesContentWidget.dart';
 import 'package:illinois/ui/academics/SkillsSelfEvaluation.dart';
 import 'package:illinois/ui/academics/StudentCourses.dart';
@@ -45,7 +46,7 @@ enum AcademicsContent { events,
   gies_checklist, uiuc_checklist,
   canvas_courses, medicine_courses, student_courses,
   skills_self_evaluation,
-  todo_list, due_date_catalog, my_illini, appointments
+  todo_list, due_date_catalog, my_illini, appointments, skills_dashboard
 }
 
 class AcademicsHomePanel extends StatefulWidget {
@@ -153,7 +154,7 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
       ),
       Expanded(child:
         Stack(children: [
-          Padding(padding: _skillsSelfEvaluationSelected ? EdgeInsets.zero : EdgeInsets.only(top: 16, left: 16, right: 16,), child:
+          Padding(padding: _skillsSelfEvaluationSelected || _skillsDashboardSelected ? EdgeInsets.zero : EdgeInsets.only(top: 16, left: 16, right: 16,), child:
             _contentWidget
           ),
           _buildContentValuesContainer()
@@ -273,6 +274,8 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
     } else if (code == 'my_illini') {
       return AcademicsContent.my_illini;
     } else if (code == 'appointments') {
+      return AcademicsContent.skills_dashboard;
+    } else if (code == 'skills_dashboard') {
       return AcademicsContent.appointments;
     } else {
       return null;
@@ -382,6 +385,8 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
     return ((_selectedContent == AcademicsContent.gies_checklist) ||
             (_selectedContent == AcademicsContent.uiuc_checklist) ||
             (_selectedContent == AcademicsContent.student_courses) ||
+            (_selectedContent == AcademicsContent.todo_list) ||
+            (_selectedContent == AcademicsContent.skills_dashboard) ||
             (_selectedContent == AcademicsContent.appointments) ||
             (_selectedContent == AcademicsContent.events)) ?
       _rawContentWidget :
@@ -411,6 +416,8 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
         return SkillsSelfEvaluation();
       case AcademicsContent.todo_list:
         return WellnessToDoHomeContentWidget();
+      case AcademicsContent.skills_dashboard:
+        return EssentialSkillsCoachDashboardPanel();
       case AcademicsContent.due_date_catalog:
         String? guideId = Guide().detailIdFromUrl(Config().dateCatalogUrl);
         return (guideId != null) ? GuideDetailWidget(key: _dueDateCatalogKey, guideEntryId: guideId, headingColor: Styles().colors.background) : Container();
@@ -422,6 +429,7 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
   }
   
   bool get _skillsSelfEvaluationSelected => _selectedContent == AcademicsContent.skills_self_evaluation;
+  bool get _skillsDashboardSelected => _selectedContent == AcademicsContent.skills_dashboard;
 
   bool _isCheckListCompleted(String contentKey) {
     int stepsCount = CheckList(contentKey).progressSteps?.length ?? 0;
@@ -455,6 +463,8 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
         return Localization().getStringEx('panel.academics.section.my_illini.label', 'myIllini');
       case AcademicsContent.appointments:
         return Localization().getStringEx('panel.academics.section.appointments.label', 'Appointments');
+      case AcademicsContent.skills_dashboard:
+        return Localization().getStringEx('', 'Essential Skills Coach');
     }
   }