+++ lib/ui/wellness/todo/WellnessToDoHomeContentWidget.dart
@@ -18,7 +18,7 @@ import 'dart:collection';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:illinois/model/wellness/ToDo.dart';
+import 'package:illinois/model/wellness/WellnessToDo.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/Wellness.dart';
@@ -29,6 +29,7 @@ import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
+import 'package:rokwire_plugin/ui/widgets/scroll_pager.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class WellnessToDoHomeContentWidget extends StatefulWidget {
@@ -42,10 +43,15 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
   static final String _unAssignedLabel =
       Localization().getStringEx('panel.wellness.todo.items.unassigned.category.label', 'Unassigned Items');
   late _ToDoTab _selectedTab;
-  List<ToDoItem>? _todoItems;
+  List<WellnessToDoItem> _todoItems = [];
+  List<WellnessToDoItem> _recurringTodoItems = [];
   late DateTime _calendarStartDate;
   late DateTime _calendarEndDate;
   bool _itemsLoading = false;
+  late ScrollPagerController _scrollPagerController;
+  final ScrollController _scrollController = ScrollController();
+
+  int _limit = 100;
 
   @override
   void initState() {
@@ -59,18 +65,33 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     ]);
     _selectedTab = _ToDoTab.daily;
     _initCalendarDates();
-    _loadToDoItems();
+    // _loadToDoItems();
+
+    _scrollPagerController = ScrollPagerController(limit: _limit, controller:_scrollController, /*infinite: true,*/ onPage: _loadTodos, onStateChanged: (){
+      if (mounted) {
+        setState(() {
+        });
+      }
+    }, onReset: () {
+      _todoItems = [];
+      _recurringTodoItems = [];
+    });
+
+    _scrollPagerController.registerScrollController(_scrollController);
+
     if (Wellness().isToDoListAccessed != true) {
       Wellness().toDoListAccessed(true);
       WidgetsBinding.instance.addPostFrameCallback((_) {
         _showWelcomePopup();
       });
     }
+
   }
 
   @override
   void dispose() {
     NotificationService().unsubscribe(this);
+    _scrollController.dispose();
     super.dispose();
   }
 
@@ -91,7 +112,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
           _buildManageButtonsRow(),
           _buildCalendarWidget(),
           _buildItemsContent(),
-          _buildClearCompletedItemsButton()
+          // _buildClearCompletedItemsButton()
         ]));
   }
 
@@ -146,7 +167,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
         ]));
   }
 
-  Widget _buildClearCompletedItemsButton() {
+  /* Unused: Widget _buildClearCompletedItemsButton() {
     return Visibility(
         visible: _clearCompletedItemsButtonVisible,
         child: Padding(
@@ -158,7 +179,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
                 contentWeight: 0.75,
                 padding: EdgeInsets.symmetric(vertical: 8),
                 onTap: _onTapClearCompletedItems)));
-  }
+  }*/
 
   Widget _buildCalendarWidget() {
     if (_selectedTab != _ToDoTab.weekly) {
@@ -256,9 +277,9 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     DateTime currentDate = DateTime.fromMillisecondsSinceEpoch(_calendarStartDate.millisecondsSinceEpoch);
     while (currentDate.isBefore(_calendarEndDate)) {
       List<Widget> dayItemWidgets = <Widget>[];
-      List<ToDoItem>? dayItems = _getItemsForDate(currentDate);
+      List<WellnessToDoItem>? dayItems = _getItemsForDate(currentDate);
       if (CollectionUtils.isNotEmpty(dayItems)) {
-        for (ToDoItem item in dayItems!) {
+        for (WellnessToDoItem item in dayItems!) {
           dayItemWidgets.add(Padding(padding: EdgeInsets.only(top: 7), child: _buildCalendarToDoItem(item)));
         }
       }
@@ -275,7 +296,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
         padding: EdgeInsets.only(top: 34), child: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: scrollWidgets));
   }
 
-  Widget _buildCalendarToDoItem(ToDoItem? item) {
+  Widget _buildCalendarToDoItem(WellnessToDoItem? item) {
     double widgetSize = 30;
     bool hasReminder = (item?.reminderDateTimeUtc != null);
     return  GestureDetector(
@@ -290,18 +311,26 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
   }
 
   Widget _buildItemsContent() {
+
     if (_sortedItemsMap != null) {
       List<Widget> contentList = <Widget>[];
       for (String key in _sortedItemsMap!.keys) {
-        List<ToDoItem>? items = _sortedItemsMap![key];
+        List<WellnessToDoItem>? items = _sortedItemsMap![key];
         if (CollectionUtils.isNotEmpty(items)) {
           contentList.add(_buildSectionWidget(key));
-          for (ToDoItem item in items!) {
+          for (WellnessToDoItem item in items!) {
             contentList.add(Padding(padding: EdgeInsets.only(top: 10), child: _ToDoItemCard(item: item)));
           }
         }
       }
-      return Column(children: contentList);
+
+      return Expanded(
+        child: ScrollPager(
+          controller: _scrollPagerController,
+          child: Column(children: contentList),
+        ),
+      );
+      //Dead code: return Column(children: contentList);
     } else {
       return _buildEmptyContent();
     }
@@ -385,7 +414,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     }
   }
 
-  void _onTapClearCompletedItems() {
+  /* Unused: void _onTapClearCompletedItems() {
     Analytics().logSelect(target: 'Clear Completed Items', source: widget.runtimeType.toString());
     if (CollectionUtils.isEmpty(_todoItems)) {
       AppAlert.showDialogResult(context, Localization().getStringEx('panel.wellness.todo.items.no_items.msg', 'There are no To-Do items.'));
@@ -396,9 +425,9 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
         message: Localization().getStringEx(
             'panel.wellness.todo.item.completed.delete.confirmation.msg', 'Are you sure that you want to delete all completed To-Do items?'),
         positiveCallback: () => _deleteCompletedItems());
-  }
+  } */
 
-  void _deleteCompletedItems() {
+  /* Unused: void _deleteCompletedItems() {
     Analytics().logWellness(
       category: Analytics.LogWellnessCategoryToDo,
       action: Analytics.LogWellnessActionClear,
@@ -424,7 +453,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
       AppAlert.showDialogResult(context, msg);
       _setItemsLoading(false);
     });
-  }
+  } */
 
   void _onTapPreviousWeek() {
     Analytics().logSelect(target: "Previous Week", source: widget.runtimeType.toString());
@@ -442,7 +471,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     _updateState();
   }
 
-  void _onTapCalendarItem(ToDoItem? item) async {
+  void _onTapCalendarItem(WellnessToDoItem? item) async {
     Analytics().logSelect(target: "Calendar", source: widget.runtimeType.toString());
     if (item == null) {
       return;
@@ -479,20 +508,11 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     _calendarEndDate = now.add(Duration(days: (7 - (now.weekday + 1))));
   }
 
-  void _loadToDoItems() {
-    _setItemsLoading(true);
-    Wellness().loadToDoItems().then((items) {
-      _todoItems = items;
-      _sortItemsByDate();
-      _setItemsLoading(false);
-    });
-  }
-
   void _sortItemsByDate() {
     if (CollectionUtils.isEmpty(_todoItems)) {
       return;
     }
-    _todoItems!.sort((ToDoItem first, ToDoItem second) {
+    _todoItems.sort((WellnessToDoItem first, WellnessToDoItem second) {
       if ((first.dueDateTime == null) && (second.dueDateTime != null)) {
         return -1;
       } else if ((first.dueDateTime != null) && (second.dueDateTime == null)) {
@@ -505,7 +525,7 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     });
   }
 
-  String? _getItemKeyByTab(ToDoItem item) {
+  String? _getItemKeyByTab(WellnessToDoItem item) {
     String? key;
     switch (_selectedTab) {
       case _ToDoTab.daily:
@@ -524,24 +544,29 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     return key;
   }
 
-  Map<String, List<ToDoItem>>? get _sortedItemsMap {
+  Map<String, List<WellnessToDoItem>>? get _sortedItemsMap {
     if (CollectionUtils.isEmpty(_todoItems)) {
       return null;
     }
-    Map<String, List<ToDoItem>> itemsMap = {};
-    for (ToDoItem item in _todoItems!) {
+
+    Map<String, List<WellnessToDoItem>> itemsMap = {};
+    List<WellnessToDoItem>? categoryItems;
+    for (WellnessToDoItem item in _todoItems) {
       String itemKey = _getItemKeyByTab(item)!;
-      List<ToDoItem>? categoryItems = itemsMap[itemKey];
+      categoryItems = itemsMap[itemKey];
       if (categoryItems == null) {
-        categoryItems = <ToDoItem>[];
+        categoryItems = <WellnessToDoItem>[];
       }
       categoryItems.add(item);
       itemsMap[itemKey] = categoryItems;
     }
+
+    _sortItemsByDate();
+
     if (_selectedTab != _ToDoTab.category) {
       return itemsMap;
     }
-    SplayTreeMap<String, List<ToDoItem>> sortedMap = SplayTreeMap<String, List<ToDoItem>>.from(itemsMap, (first, second) {
+    SplayTreeMap<String, List<WellnessToDoItem>> sortedMap = SplayTreeMap<String, List<WellnessToDoItem>>.from(itemsMap, (first, second) {
       if ((first == _unAssignedLabel) && (second != _unAssignedLabel)) {
         return -1;
       } else if ((first != _unAssignedLabel) && (second == _unAssignedLabel)) {
@@ -553,28 +578,26 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     return sortedMap;
   }
 
-  List<ToDoItem>? _getItemsForDate(DateTime date) {
-    List<ToDoItem>? dayItems;
-    if (_todoItems != null) {
-      for (ToDoItem item in _todoItems!) {
-        DateTime? itemDueDate = item.dueDateTime;
-        if (itemDueDate != null) {
-          if ((itemDueDate.year == date.year) && (itemDueDate.month == date.month) && (itemDueDate.day == date.day)) {
-            if (dayItems == null) {
-              dayItems = <ToDoItem>[];
-            }
-            dayItems.add(item);
+  List<WellnessToDoItem>? _getItemsForDate(DateTime date) {
+    List<WellnessToDoItem>? dayItems;
+    for (WellnessToDoItem item in _todoItems) {
+      DateTime? itemDueDate = item.dueDateTime;
+      if (itemDueDate != null) {
+        if ((itemDueDate.year == date.year) && (itemDueDate.month == date.month) && (itemDueDate.day == date.day)) {
+          if (dayItems == null) {
+            dayItems = <WellnessToDoItem>[];
           }
+          dayItems.add(item);
         }
       }
     }
     return dayItems;
   }
 
-  void _setItemsLoading(bool loading) {
+  /* Unused void _setItemsLoading(bool loading) {
     _itemsLoading = loading;
     _updateState();
-  }
+  } */
 
   void _updateState() {
     if (mounted) {
@@ -582,35 +605,35 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
     }
   }
 
-  List<String>? get _completedItemIds {
+  /* Ununsed List<String>? get _completedItemIds {
     if (CollectionUtils.isEmpty(_todoItems)) {
       return null;
     }
     List<String> completedItemsIds = <String>[];
-    for (ToDoItem item in _todoItems!) {
+    for (ToDoItem item in _todoItems) {
       if (item.isCompleted) {
         completedItemsIds.add(item.id!);
       }
     }
     return completedItemsIds;
-  }
+  } */
 
-  List<String>? get _completedItemNames {
+  /* Ununsed List<String>? get _completedItemNames {
     if (CollectionUtils.isEmpty(_todoItems)) {
       return null;
     }
     List<String> completedItemsNames = <String>[];
-    for (ToDoItem item in _todoItems!) {
+    for (ToDoItem item in _todoItems) {
       if (item.isCompleted) {
         completedItemsNames.add(item.name!);
       }
     }
     return completedItemsNames;
-  }
+  } */
 
-  bool get _clearCompletedItemsButtonVisible {
+  /* Ununsed bool get _clearCompletedItemsButtonVisible {
     return (_completedItemIds?.length ?? 0) > 0;
-  }
+  } */
 
   String get _formattedCalendarMonthLabel {
     if (_calendarStartDate.month != _calendarEndDate.month) {
@@ -631,13 +654,170 @@ class _WellnessToDoHomeContentWidgetState extends State<WellnessToDoHomeContentW
         (name == Wellness.notifyToDoItemsDeleted) ||
         (name == Wellness.notifyToDoCategoryChanged) ||
         (name == Wellness.notifyToDoCategoryDeleted)) {
-      _loadToDoItems();
+        _scrollPagerController.reset();
+
     }
   }
+
+  Future<int> _loadTodos({required int limit, required int offset}) async {
+    List<WellnessToDoItem>? responses = await Wellness().loadToDoItems(limit, offset);
+    if (mounted && responses != null) {
+      setState(() {
+        _todoItems.addAll(responses.where((a) => _todoItems.every((b) => a.id != b.id)));
+        // _todoItems?.addAll(responses);
+        _recurringTodoItems.addAll(_todoItems.where((element) => element.recurrenceType != null && element.recurrenceType != "none" && element.recurrenceId == null));
+        _fillRecurringTodos();
+      });
+    }
+    return responses?.length ?? -1;
+  }
+
+  void _fillRecurringTodos(){
+    if(_recurringTodoItems.isNotEmpty){
+      DateTime? date = _recurringTodoItems[0].dueDateTimeUtc ?? DateTime.now();
+      List<WellnessToDoItem> recurringItems = [];
+      while(_todoItems.length + recurringItems.length < _limit){
+        for(WellnessToDoItem item in _recurringTodoItems){
+          if(date!.isBefore(item.dueDateTimeUtc ?? DateTime.now())
+              || _isSameDate(item.dueDateTimeUtc, date)
+              || _isRecurringDayCompleted(item, date)
+          ){
+            continue;
+          }else{
+            //daily
+            if(item.recurrenceType?.substring((item.recurrenceType?.length ?? 0) -5, (item.recurrenceType?.length ?? 0)) == "* * *"){
+              WellnessToDoItem toDoItem = WellnessToDoItem(
+                  name: item.name,
+                  category: item.category,
+                  dueDateTimeUtc: new DateTime.utc(date.year, date.month, date.day, item.dueDateTimeUtc?.hour ?? 0, item.dueDateTimeUtc?.minute ?? 0),
+                  hasDueTime:  item.hasDueTime,
+                  reminderType:  item.reminderType,
+                  reminderDateTimeUtc: item.reminderDateTimeUtc,
+                  workDays:  item.workDays,
+                  location:  item.location,
+                  description:  item.description,
+                  isCompleted: false,
+                  recurrenceType:  item.recurrenceType,
+                  recurrenceId:  item.id
+              );
+              recurringItems.add(toDoItem);
+              //monthly
+            }else if(item.recurrenceType?.substring((item.recurrenceType?.length ?? 0) -5, (item.recurrenceType?.length ?? 0)) == "* ? *"){
+              List<String>? result = item.recurrenceType?.split(' ');
+              if(result != null && int.parse(result[3]) == date.day) {
+                WellnessToDoItem toDoItem = WellnessToDoItem(
+                    name: item.name,
+                    category: item.category,
+                    dueDateTimeUtc: new DateTime.utc(date.year, date.month, date.day, item.dueDateTimeUtc?.hour ?? 0, item.dueDateTimeUtc?.minute ?? 0),
+                    hasDueTime:  item.hasDueTime,
+                    reminderType:  item.reminderType,
+                    reminderDateTimeUtc: item.reminderDateTimeUtc,
+                    workDays:  item.workDays,
+                    location:  item.location,
+                    description:  item.description,
+                    isCompleted: false,
+                    recurrenceType:  item.recurrenceType,
+                    recurrenceId:  item.id
+                );
+                recurringItems.add(toDoItem);
+              }
+            }else{
+              List<String>? result = item.recurrenceType?.split(' ');
+              if(result != null){
+                //weekly
+                if(!result[5].contains(",") && _convertDayString(result[5]) == date.weekday){
+                  WellnessToDoItem toDoItem = WellnessToDoItem(
+                      name: item.name,
+                      category: item.category,
+                      dueDateTimeUtc: new DateTime.utc(date.year, date.month, date.day, item.dueDateTimeUtc?.hour ?? 0, item.dueDateTimeUtc?.minute ?? 0),
+                      hasDueTime:  item.hasDueTime,
+                      reminderType:  item.reminderType,
+                      reminderDateTimeUtc: item.reminderDateTimeUtc,
+                      workDays:  item.workDays,
+                      location:  item.location,
+                      description:  item.description,
+                      isCompleted: false,
+                      recurrenceType:  item.recurrenceType,
+                      recurrenceId:  item.id
+                  );
+                  recurringItems.add(toDoItem);
+
+                  //weekdays
+                }else{
+                  List<String>? daysStringList  = item.recurrenceType?.split(" ");
+                  List<String>? dayStrings = daysStringList?[5].split(",");
+                  if(dayStrings != null){
+                    for(String day in dayStrings){
+                      if(_convertDayString(day) == date.weekday){
+                        WellnessToDoItem toDoItem = WellnessToDoItem(
+                            name: item.name,
+                            category: item.category,
+                            dueDateTimeUtc: new DateTime.utc(date.year, date.month, date.day, item.dueDateTimeUtc?.hour ?? 0, item.dueDateTimeUtc?.minute ?? 0),
+                            hasDueTime:  item.hasDueTime,
+                            reminderType:  item.reminderType,
+                            reminderDateTimeUtc: item.reminderDateTimeUtc,
+                            workDays:  item.workDays,
+                            location:  item.location,
+                            description:  item.description,
+                            isCompleted: false,
+                            recurrenceType:  item.recurrenceType,
+                            recurrenceId:  item.id
+                        );
+                        recurringItems.add(toDoItem);
+                        break;
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+        date = date?.add(Duration(days: 1));
+      }
+      _todoItems.addAll(recurringItems);
+    }
+  }
+
+  int _convertDayString(String day){
+    switch(day){
+      case "MON":
+        return 1;
+      case "TUE":
+        return 2;
+      case "WED":
+        return 3;
+      case "THU":
+        return 4;
+      case "FRI":
+        return 5;
+      case "SAT":
+        return 6;
+      case "SUN":
+        return 7;
+      default:
+        return -1;
+
+    }
+  }
+
+  bool _isSameDate(DateTime? item, DateTime? date){
+    return item?.month == date?.month && item?.day == date?.day;
+  }
+
+  bool _isRecurringDayCompleted(WellnessToDoItem recurringItem, DateTime? date){
+    for(WellnessToDoItem item in _todoItems){
+      if(_isSameDate(item.dueDateTimeUtc, date) && (item.recurrenceId ?? "") == recurringItem.id){
+        return true;
+      }
+    }
+    return false;
+  }
+
 }
 
 class _ToDoItemCard extends StatefulWidget {
-  final ToDoItem item;
+  final WellnessToDoItem item;
   _ToDoItemCard({required this.item});
 
   @override
@@ -683,18 +863,30 @@ class _ToDoItemCardState extends State<_ToDoItemCard> {
     widget.item.isCompleted = !widget.item.isCompleted;
     AppSemantics.announceCheckBoxStateChange(context, widget.item.isCompleted , widget.item.name);
     _setLoading(true);
-    Wellness().updateToDoItem(widget.item).then((success) {
-      if (!success) {
-        // revert value if update fails
-        widget.item.isCompleted = !widget.item.isCompleted;
-        String msg = Localization().getStringEx('panel.wellness.todo.item.update.failed.msg', 'Failed to update To-Do item.');
-        AppAlert.showDialogResult(context, msg);
-      }
-      _setLoading(false);
-    });
+    if(widget.item.id != null){
+      Wellness().updateToDoItem(widget.item).then((success) {
+        if (!success) {
+          // revert value if update fails
+          widget.item.isCompleted = !widget.item.isCompleted;
+          String msg = Localization().getStringEx('panel.wellness.todo.item.update.failed.msg', 'Failed to update To-Do item.');
+          AppAlert.showDialogResult(context, msg);
+        }
+
+      });
+    }else{
+      Wellness().createToDoItem(widget.item).then((success) {
+        if (!success) {
+          // revert value if update fails
+          widget.item.isCompleted = !widget.item.isCompleted;
+          String msg = Localization().getStringEx('panel.wellness.todo.item.create.failed.msg', 'Failed to create To-Do item.');
+          AppAlert.showDialogResult(context, msg);
+        }
+      });
+    }
+    _setLoading(false);
   }
 
-  void _onTapEdit(ToDoItem item) {
+  void _onTapEdit(WellnessToDoItem item) {
     Analytics().logSelect(target: "Edit Item", source: widget.runtimeType.toString());
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessToDoItemDetailPanel(item: item)));
   }
@@ -766,7 +958,7 @@ class _TabButton extends StatelessWidget {
 }
 
 class _ToDoItemReminderDialog extends StatefulWidget {
-  final ToDoItem item;
+  final WellnessToDoItem item;
   _ToDoItemReminderDialog({required this.item});
 
   @override
@@ -775,7 +967,7 @@ class _ToDoItemReminderDialog extends StatefulWidget {
 }
 
 class _ToDoItemReminderDialogState extends State<_ToDoItemReminderDialog> {
-  late ToDoItem _item;
+  late WellnessToDoItem _item;
   late DateTime _reminderDateTime;
   bool _loading = false;
 
@@ -889,7 +1081,7 @@ class _ToDoItemReminderDialogState extends State<_ToDoItemReminderDialog> {
     );
     _setLoading(true);
     _item.reminderDateTimeUtc = _reminderDateTime.toUtc();
-    _item.reminderType = ToDoReminderType.specific_time;
+    _item.reminderType = WellnessToDoReminderType.specific_time;
     Wellness().updateToDoItem(_item).then((success) {
       _setLoading(false);
       if (!success) {
@@ -927,4 +1119,6 @@ class _ToDoItemReminderDialogState extends State<_ToDoItemReminderDialog> {
   String get _formattedTime {
     return AppDateTime().formatDateTime(_reminderDateTime, format: 'hh : mm a', ignoreTimeZone: true)!;
   }
+
+
 }